Creating Caesar Chiper
Caesar Cipher Technique is a simple and easy method of encryption technique. Each letter of plain text is replaced by a letter with some fixed number of positions 
down with the alphabet. e.g. every letter of the message is replaced by it's nearest consequent(A becomes B, B becomes C...). The only exception is Z which becomes A.